fst: FmdAttr: set a dud SetSyncStatus impl
Since this function is primarily used for GetInconsistency, we are still not
using this yet. So far the FmdAttr class has no data members and is technically
not needing a global constr./destructor. This might change as we implement the
GetInconsistencyStats method and whether we need a Fs Level lock or not.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>